% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{cairo_failsafe}
\alias{cairo_failsafe}
\title{Try to use \code{cairo_pdf}, but fall back to \code{pdf} if that doesn't work.}
\usage{
cairo_failsafe(...)
}
\arguments{
\item{...}{arguments to cairo_pdf/pdf}
}
\description{
Try to use \code{cairo_pdf}, but fall back to \code{pdf} if that doesn't work.
}

