% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{ExtractSignificantResults}
\alias{ExtractSignificantResults}
\title{Extracts the significant results from TestMatchedControlMixedModel and
adjusts for multiple testing (p-values corrected are only those for the
interaction terms with the correct magnitude sign)}
\usage{
ExtractSignificantResults(glmm.results, direction = `>`,
  padj.threshold = 0.1)
}
\arguments{
\item{glmm.results}{results from TestMatchedControlMixedModel}

\item{direction}{a binary function such as > or < determining coefficient relationship to 0}

\item{padj.threshold}{only return adjusted p-values below this value}
}
\description{
Extracts the significant results from TestMatchedControlMixedModel and
adjusts for multiple testing (p-values corrected are only those for the
interaction terms with the correct magnitude sign)
}

