% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{fct_anon_deterministic}
\alias{fct_anon_deterministic}
\title{Non-random version of \code{\link{forcats::fct_anon}}}
\usage{
fct_anon_deterministic(f, prefix = "")
}
\arguments{
\item{f}{A factor}

\item{prefix}{A character prefix to insert in front of the random labels.}
}
\description{
Non-random version of \code{\link{forcats::fct_anon}}
}

