% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{LinkMatchedControls}
\alias{LinkMatchedControls}
\title{Links matched samples and (environmental) controls via the grouping.col
by completing any missing cases.}
\usage{
LinkMatchedControls(agg, s, md, grouping.col)
}
\arguments{
\item{agg}{agglomerated dataframe}

\item{s}{sample metadata}

\item{md}{otu metadata}
}
\value{
the modified agg with a new column, PlotGroup, for line drawing
}
\description{
Links matched samples and (environmental) controls via the grouping.col
by completing any missing cases.
}

