% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load datasets.}
\usage{
load_data(set = c("A", "B", "C", "DE"), data = c("16S", "ITS", "virome",
  "WGS"), data_fp = opts$data_fp)
}
\arguments{
\item{set}{the sample set in question (of A-C, DE)}

\item{data}{the data type (of 16S, ITS, virome, or WGS)}

\item{data_fp}{path to data files}
}
\description{
Load datasets.
}

