% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{LoadData}
\alias{LoadData}
\title{Load datasets.}
\usage{
LoadData(set = c("A", "B", "C", "DE"), data = c("16S", "ITS", "virome",
  "WGS"), data_fp = opts$get("data_fp"))
}
\arguments{
\item{set}{the sample set in question (of A-C, DE)}

\item{data}{the data type (of 16S, ITS, virome, or WGS)}
}
\description{
Load datasets.
}

