% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{MakeBarchartData}
\alias{MakeBarchartData}
\title{Groups taxa into top members and samples by their larger group (if any)
in preparation for plotting summary barcharts}
\usage{
MakeBarchartData(agg, grouping.col = "SubjectID", taxa.level.1 = "Order",
  taxa.level.2 = "Order", top.1 = 8, top.2 = 0)
}
\arguments{
\item{agg}{an agglomerated dataframe}

\item{taxa.level.1}{the primary taxonomic rank to show in the barcharts}

\item{taxa.level.2}{the second, higher rank to show}

\item{top.1}{how many colors/levels to display}

\item{top.2}{how many to display for what didn't fit into top.1}

\item{grouping.column}{a column that defines larger groups of samples (optional; set to NA to omit)}
}
\description{
Groups taxa into top members and samples by their larger group (if any)
in preparation for plotting summary barcharts
}

