% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{MakeEqualSizePlots}
\alias{MakeEqualSizePlots}
\title{Facets on a layout matrix to make individual plot sizes consistent. Removes
the legend; grab and display it separately with get_ggplot_legend.}
\usage{
MakeEqualSizePlots(p, data, .variables = c("StudyGroup", "SubjectID"),
  at_most = 45, ncol = 1, extra_text = "")
}
\arguments{
\item{p}{a ggplot object (whose data will be replaced by the chunked data)}

\item{data}{the dataframe to facet}

\item{.variables}{a character vector listing the columns to facet by in data}

\item{at_most}{make space for at most this many facets}

\item{ncol}{the number of columns in the layout matrix (must be multiple of at_most)}
}
\description{
Facets on a layout matrix to make individual plot sizes consistent. Removes
the legend; grab and display it separately with get_ggplot_legend.
}

