% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{MakeHeatmapData}
\alias{MakeHeatmapData}
\title{Prepare agglomerated dataframe for heatmap display by aggregating counts
by minimum rank, completing missing cases, and clustering rows}
\usage{
MakeHeatmapData(agg, s, min.rank.1 = "Genus", min.rank.2 = NA,
  min.samples = 1)
}
\arguments{
\item{s}{sample metadata dataframe}

\item{min.rank.1}{the most specific rank to aggregate by (determines y-axis)
Uses tax_climber to select next most specific rank if not available.}

\item{min.rank.2}{(optional) higher rank to prefix min.rank.1 (e.g. Phylum)
Uses tax_climber to select next most specific rank if not available.}

\item{min.samples}{the minimum number of samples required for a taxa to appear}

\item{min.samples}{a row must appear more than this number of samples to be shown}
}
\description{
Prepare agglomerated dataframe for heatmap display by aggregating counts
by minimum rank, completing missing cases, and clustering rows
}

