% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{MakePCOAData}
\alias{MakePCOAData}
\title{Performs principle coordinates analysis on a GUniFrac distance and merges
sample metadata for plotting.}
\usage{
MakePCOAData(unifrac.data, s, grouping.col)
}
\arguments{
\item{unifrac.data}{the output of MakeUnifracData}

\item{s}{sample metadata dataframe}

\item{grouping.col}{column in s defining centroid groups}
}
\description{
Performs principle coordinates analysis on a GUniFrac distance and merges
sample metadata for plotting.
}

