% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{MakeTopBarchartData}
\alias{MakeTopBarchartData}
\title{Prepares data for making the summary barcharts showing the dominant taxa
by group (rather than by sample).}
\usage{
MakeTopBarchartData(agg, s, top = 6, min.rank = "Species")
}
\arguments{
\item{agg}{agglomerated data frame, filtered as desired}

\item{s}{sample metadata}

\item{top}{top taxa to select from each group (displayed taxa will be union)}

\item{min.rank}{min taxonomic rank to display (will group at this or higher)}
}
\description{
Prepares data for making the summary barcharts showing the dominant taxa
by group (rather than by sample).
}

