% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa.R
\name{MakeUnifracData}
\alias{MakeUnifracData}
\title{Calculates the generalized UniFrac distances for an agglomerated dataframe.
This involves a random rooting of the tree and rarefaction, which are
non-deterministic. For reproducibility, recommend calling set.seed before this}
\usage{
MakeUnifracData(agg, s, tree, rarefy.depth, alpha = 0.5)
}
\arguments{
\item{agg}{agglomerated dataframe}

\item{s}{sample metadata dataframe}

\item{tree}{a tree (will root using random node)}

\item{alpha}{the alpha parameter for GUniFrac}
}
\description{
Calculates the generalized UniFrac distances for an agglomerated dataframe.
This involves a random rooting of the tree and rarefaction, which are
non-deterministic. For reproducibility, recommend calling set.seed before this
}

