% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{min_rank}
\alias{min_rank}
\title{Creates a MinRank vector from an agglomerated data frame (based on
eclectic::tax_climber). Use inside dplyr::mutate or directly assign to col.
MinRank columns are the most specific taxonomic assignments (up to a
threshold)}
\usage{
min_rank(agg, end.rank, ranks = qiimer::taxonomic_ranks, ...)
}
\arguments{
\item{agg}{agglomerated data frame (with taxonomic ranks)}

\item{end.rank}{the most specific taxonomic rank if available}

\item{...}{additional arguments passed to eclectic::tax_climber}

\item{rank}{a character vector of taxonomic ranks present in agg}
}
\description{
Creates a MinRank vector from an agglomerated data frame (based on
eclectic::tax_climber). Use inside dplyr::mutate or directly assign to col.
MinRank columns are the most specific taxonomic assignments (up to a
threshold)
}

