% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\docType{data}
\name{opts}
\alias{opts}
\title{Get and set global options.}
\format{An object of class \code{list} of length 2.}
\usage{
opts
}
\description{
Get and set global options.
}
\keyword{datasets}

