% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa.R
\name{PCoAChunk}
\alias{PCoAChunk}
\title{Wrapper for UniFrac calculations, PERMANOVA testing and PCoA plotting.}
\usage{
PCoAChunk(agg, s, tree, rarefy.depth, testing.terms, sampleset = c("A", "B",
  "C"), dataset = c("16S", "ITS"), subset, fig.num,
  fig.dir = opts$figure_fp)
}
\description{
Wrapper for UniFrac calculations, PERMANOVA testing and PCoA plotting.
}

