% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{PlotBarchart}
\alias{PlotBarchart}
\title{Plots the results of MakeBarchartData as a set of grouped barcharts.}
\usage{
PlotBarchart(.data, x.col = "SampleType", fill.col = "taxa",
  grouping.col = "SubjectID", x.labs = attributes(.data)$labels,
  fill.values = attributes(.data)$fill,
  fill.breaks = rev(levels(.data$taxa)), theme = opts$barchart.theme,
  guide = opts$barchart.guide)
}
\arguments{
\item{.data}{result from \code{\link{MakeBarchartData}}}

\item{x.col}{the mapping for the x-axis}

\item{fill.col}{the mapping for the fill aesthetic}

\item{grouping.col}{the grouping columns specified in MakeBarchartData, if any}

\item{x.labs}{if desired, what to revalue the x-axis labels to}

\item{fill.values}{named vector for fill colors}

\item{fill.breaks}{ordering of the fills (currently a noop)}

\item{theme}{a theme object to add}
}
\value{
a \code{\link{ggplot2::ggplot}} object
}
\description{
Plots the results of MakeBarchartData as a set of grouped barcharts.
}

