% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{PlotHeatmap}
\alias{PlotHeatmap}
\title{Plot the results of MakeHeatmapData.}
\usage{
PlotHeatmap(heatmap.data, use.reads = TRUE, threshold = 0.6,
  x.text = TRUE, x.axis = "SampleID")
}
\arguments{
\item{heatmap.data}{the dataframe from MakeHeatmapData}

\item{use.reads}{if TRUE, plot the raw readcounts; if FALSE, use proportions}

\item{threshold}{at what relative proportion the color scale goes to red}

\item{x.text}{display labels on the x-axis}
}
\description{
Plot the results of MakeHeatmapData.
}

