% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{PlotPCOA}
\alias{PlotPCOA}
\title{Standardized PCoA plot.}
\usage{
PlotPCOA(pcoa.data, fill = "StudyGroup", linetype = "SampleType",
  fill.palette, color.palette)
}
\arguments{
\item{pcoa.data}{dataframe generated by MakePCOAData}

\item{fill}{column to set as fill aesthetic}

\item{linetype}{column to set as linetype aesthetic}

\item{fill.palette}{color palette to use as fill}

\item{color.palette}{color palette to use for outlines and line segments}
}
\description{
Standardized PCoA plot.
}

