% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{PlotSingleBarchart}
\alias{PlotSingleBarchart}
\title{Based on PlotBarchart, but facets along ID.col. Intended for use with
MakeEqualSizePlots function. Not suggested for use on its own.}
\usage{
PlotSingleBarchart(..., fill.values, fill.breaks = waiver(),
  ID.col = "SubjectID")
}
\arguments{
\item{fill.values}{the palette to use for the scale_fill_manual}

\item{ID.col}{the column with the desired display indices}
}
\description{
Based on PlotBarchart, but facets along ID.col. Intended for use with
MakeEqualSizePlots function. Not suggested for use on its own.
}

