% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{PlotTopBarcharts}
\alias{PlotTopBarcharts}
\title{Plots data prepared by \link{MakeTopBarchartData}.}
\usage{
PlotTopBarcharts(top.data)
}
\description{
Plots data prepared by \link{MakeTopBarchartData}.
}

