% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{root_tree}
\alias{root_tree}
\title{Root a tree using a random node as the root.}
\usage{
root_tree(tree, otus)
}
\arguments{
\item{tree}{a phylogenetic tree (class "phylo")}

\item{otus}{vector of otus to keep}
}
\value{
rooted tree
}
\description{
Root a tree using a random node as the root.
}

