% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{RunDESeq2}
\alias{RunDESeq2}
\title{Test differences between study groups using DESeq2.}
\usage{
RunDESeq2(ranks.to.test, agg, md, sarcoid.name = "sarcoid")
}
\arguments{
\item{ranks.to.test}{a vector giving the taxonomic ranks to collapse and test}

\item{agg}{an appropriately filtered agglomerated data frame}

\item{md}{the metadata dataframe}
}
\description{
Test differences between study groups using DESeq2.
}

