% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{RunGLMM}
\alias{RunGLMM}
\title{Test differences between study groups using the generalized linear mixed model}
\usage{
RunGLMM(ranks.to.test, agg, md, sparsity = 0.1,
  grouping.var = "OriginatingSampleID")
}
\arguments{
\item{ranks.to.test}{a vector giving the taxonomic ranks to collapse and test}

\item{agg}{an appropriately filtered agglomerated data frame}

\item{md}{the metadata dataframe}

\item{sparsity}{omit testing taxa that appear in fewer that this pct of samples}
}
\description{
Test differences between study groups using the generalized linear mixed model
}

