% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{StandardizeMetadata}
\alias{StandardizeMetadata}
\title{Standardize a metadata table (s) to be compatible with SRA_data requirements}
\usage{
StandardizeMetadata(ds)
}
\arguments{
\item{ds}{a SarcoidDataset object created by \code{\link{LoadData}}}
}
\description{
Standardize a metadata table (s) to be compatible with SRA_data requirements
}

