% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{TestMatchedControlMixedModel}
\alias{TestMatchedControlMixedModel}
\title{Create and test a mixed effect model between two sample types and study groups
(e.g. BAL and prewash, tissue and paraffin), where the interaction term for the
model is the coefficient of interest.}
\usage{
TestMatchedControlMixedModel(agg, grouping.var = "OriginatingSampleID",
  tax.level = "Genus", sparsity = 0.1)
}
\arguments{
\item{agg}{an agglomerated data frame; ensure the StudyGroup and SampleType
levels are ordered correctly (i.e. interesting ones last)}

\item{grouping.var}{the column defining the sample/control grouping}

\item{tax.level}{the taxonomic level to test; lower levels will be summed}

\item{sparsity}{the proportions of zeros below which a taxa will be skipped}
}
\description{
Create and test a mixed effect model between two sample types and study groups
(e.g. BAL and prewash, tissue and paraffin), where the interaction term for the
model is the coefficient of interest.
}

