% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{TestPermanova}
\alias{TestPermanova}
\title{Performs PERMANOVA on the given terms using the unifrac distance matrix.}
\usage{
TestPermanova(unifrac.data, s, terms)
}
\arguments{
\item{unifrac.data}{matrix/dist of unifrac distances (coerced to dist if not)}

\item{s}{sample metadata}

\item{terms}{a string giving the RHS of the adonis formula (eg "counts*StudyGroup")}
}
\description{
Performs PERMANOVA on the given terms using the unifrac distance matrix.
}

