% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{MakeBarchartData}
\alias{MakeBarchartData}
\title{Groups taxa into top members and samples by their larger group (if any)
in preparation for plotting summary barcharts}
\usage{
MakeBarchartData(agg, sampleset, datatype, grouping.col = "SubjectID",
  taxa.level = "Order", top.n = 8)
}
\arguments{
\item{agg}{an agglomerated dataframe (\code{\link{eclectic::agglomerate}})}

\item{taxa.level}{the taxonomic rank to show in the barcharts}

\item{top.n}{how many colors/levels to display}

\item{grouping.column}{a column that defines larger groups of samples (optional; set to NA to omit)}
}
\description{
Groups taxa into top members and samples by their larger group (if any)
in preparation for plotting summary barcharts
}

