% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarcoid_manuscript.R
\name{PlotBarchart}
\alias{PlotBarchart}
\title{Plots the results of MakeBarchartData as a set of grouped barcharts.}
\usage{
PlotBarchart(.data, x.axis.col = "SampleType", fill.col = "taxa", x.labels,
  grouping.col, reorder.taxa = TRUE)
}
\arguments{
\item{.data}{data.frame from MakeBarchartData}

\item{x.axis.col}{the mapping for the x-axis}

\item{fill.col}{the mapping for the fill aesthetic}

\item{x.labels}{if desired, what to revalue the x-axis labels to}

\item{grouping.col}{the grouping columns specified in MakeBarchartData, if any}

\item{reorder.taxa}{a noop (do it outside this function)}
}
\value{
a ggplot object
}
\description{
Plots the results of MakeBarchartData as a set of grouped barcharts.
}

